improve and correct tokenData code examples
# Getting mint number

Projects are already separated by 1000000, we don't need to also multiply it by `projectNumber`.

# Generating token data

Replaced it with more canonical code which also generates a random `tokenId`.

# Converting hash to int

Fixed the example, which was incorrect:

1. hash starts with `0x`, we don't need to include that
2. 16 chars of hex is a 64bit int, JS only supports 32bit int safely, rounding behavior of parsing numbers >52bits can and _has_ caused inconsistencies between browsers on AB projects

# Generating hex pairs

We don't need the seed and `parseInt(x, 16) % 10` isn't uniform. I'm not sure how useful it is to get integers 0-10 from hex pairs either. I think the intention was to get continuous values for use as random variables, so I updated the example to be 0-100 inclusive since percentages are intuitive.