Simplified `route` configuration (#40)
* Upgrade to docs powered by Retype

* Remove _SUMMARY.md

This file is not required with Retype as navigation is built based on the folder structure and easier to manage page or folder configurations.

* Revise icon

* Add links and formatting

* Formatting

* Update workflow config

* Use `base` config to create required repo subfolder when using GitHub Pages

* Update `url` config to use GitHub Pages public host

* Restructure pages to exactly match existing website paths

If published, this we would be a drop in replacement for the existing docs.artblocks.io website. All inbound links to docs.artblocks.io would continue to resolve correctly as the paths have not changed.

* Normalize descriptions

* Use docs.artblocks.io url

* Add basic .gitignore to just ensure the generated .retype folder is not committed

The `.retype` folder might be created when building the project locally. This folder is not required to be committed to the repo.

* Explicitly point to the readme.md

Not really required, but might help with Retype link resolution and reporting.

* This `/creator-docs` folder path is not required now

* I thought I changed this already?

Not sure why this old `repo` path is back, but updating to the correct path again.

* Correcting a link path

* Add `route` to support legacy URL's

This will ensure the new site will maintain the same URL's as the old site and any inbound links will continue to resolve correctly. No 404's from existing inbound links in the wild should be thrown.

* Rename `PBAB 101` folder to `pbab-101`

* Simplify the route config

The `route` configuration is much cleaner now.

* Support the path docs.artblocks.io/creator-docs/

* Some pages did not get updated

* Future-proof for Retype v2.0

In Retype v2.0, redirect pages will be rendered to the menu. Explicitly setting `visibility: hidden` will be required if you do not want those pages to be added to the main left navigation tree.